+++ frontend/src/components/progress/hooks/useStudyingForm.ts
